🔄 Refactor bot_api import: move to S_API directory and update references in multiple page modules
