+++ wiki_page.py
@@ -27,7 +27,7 @@ from newapi.wiki_page import MainPage, NEW_API
 # ---
 import os
 # import sys
-from newapi.super import bot_api as bot_api1
+from newapi.super.S_API import bot_api as bot_api1
 from newapi.super.S_Page import super_page as super_page1
 from newapi.super.S_Category import catdepth_new as catdepth_new1
